Fixed one more distutils import
