+++ src/wheel/wheelfile.py
@@ -5,7 +5,6 @@ import re
 import stat
 import time
 from collections import OrderedDict
-from distutils import log as logger
 from io import StringIO, TextIOWrapper
 from zipfile import ZIP_DEFLATED, ZipFile, ZipInfo
 
@@ -13,6 +12,7 @@ from wheel.cli import WheelError
 from wheel.util import (
     as_bytes,
     as_unicode,
+    log,
     native,
     urlsafe_b64decode,
     urlsafe_b64encode,
@@ -126,7 +126,7 @@ class WheelFile(ZipFile):
         return ef
 
     def write_files(self, base_dir):
-        logger.info("creating '%s' and adding '%s' to it", self.filename, base_dir)
+        log(f"creating '{self.filename}' and adding '{base_dir}' to it")
         deferred = []
         for root, dirnames, filenames in os.walk(base_dir):
             # Sort the directory names so that `os.walk` will walk them in a
@@ -166,7 +166,7 @@ class WheelFile(ZipFile):
             if isinstance(zinfo_or_arcname, ZipInfo)
             else zinfo_or_arcname
         )
-        logger.info("adding '%s'", fname)
+        log(f"adding '{fname}'")
         if fname != self.record_path:
             hash_ = self._default_algorithm(bytes)
             self._file_hashes[fname] = hash_.name, native(
